+++ utils/automation/__init__.py
@@ -1,0 +1,1 @@
+# utils.automation - Automation utilities